Add tests for setting `tunnel` option
